--- src/main/java/modelclass/Pessoa.java
@@ -1,32 +1,0 @@
-package modelclass;
-
-public class Pessoa {
-    protected String nome;
-    protected String email;
-    protected String codigo;
-
-    public Pessoa(String nome, String email) {
-        this.nome = nome;
-        this.email = email;
-    }
-
-    public String toString() {
-        return getDescricao();
-    }
-
-    public String getNome() {
-        return this.nome;
-    }
-
-    public String getEmail() {
-        return this.email;
-    }
-
-    public String getCodigo() {
-        return this.codigo;
-    }
-
-    public String getDescricao() {
-        return this.nome + " -Email:" + this.email;
-    }
-}
