+++ src/modelclass/DentistaComissionado.java
@@ -28,6 +28,7 @@ public class DentistaComissionado extends Dentista {
     public String getDentista() {
         return "É um dentista comissionado";
     }
+
     public double getPorcentagemComissao() {
         return this.porcentagemComissao;
     }
